refactor: split 'team member fields' into its own component
