+++ src/components/itineraries/Generate/Generate.tsx
@@ -9,14 +9,11 @@ import useGlobalState from "../../../hooks/useGlobalState"
 import styled from "styled-components"
 import isWeekDay from "../../../lib/utils/isWeekDay"
 import StartAddress from "../add-start-address/StartAddress"
-import UserDropdown from "./UserDropDown"
 import Modals, {caseTo, openModalTo} from "./Modals"
 import {findByProperty} from "../../../lib/utils/findByProperty"
 import {filterNullish} from "../../../lib/utils/filterNullish"
+import TeamFields from "../TeamFields"
 
-const Label = styled.label`
-  font-weight: bold
-`
 const Label2 = styled.label`
   font-weight: bold
   margin-right: 36px
@@ -34,12 +31,6 @@ const ButtonWrap = styled.div`
   display: flex
   justify-content: flex-end
 `
-
-const USER_DROPDOWNS = [
-  'Toezichthouder 1',
-  'Toezichthouder 2',
-  'Handhaver',
-]
 
 type DayPart = "day" | "evening"
 type FormValues = {
@@ -113,19 +104,7 @@ const Generate: FC = () => {
 
           return (
             <form onSubmit={handleSubmit}>
-              { USER_DROPDOWNS.map((label, index) => (
-                <Div key={index}>
-                  <Label>{ label }</Label>
-                  <Field name={`users[${index}]`}>
-                    { ({input: { onChange, value }}) => (
-                      <UserDropdown users={users ?? []}
-                        onChange={onChange}
-                        value={value}
-                        excludedUsers={values.users}
-                      />)}
-                  </Field>
-                </Div>
-              )) }
+              <TeamFields users={users ?? []} alreadySelectedUserIds={values.users}/>
               <Div>
                 <p>Wat voor looplijst wil je maken?</p>
                 { showWeekDay
